requirements.txt
colormath==3.0.0
Flask==1.1.2
Flask-Classful @ git+https://github.com/teracyhq/flask-classful.git@3bbab31705b4aa2903e7e62aa8c5ee70a1e6d789
imutils==0.5.3
matplotlib==3.3.3
numpy==1.19.5
opencv-python==4.5.1.48
paho-mqtt==1.5.1
paramiko==2.7.2
pascal-voc-writer==0.1.4
requests==2.25.1
retrying==1.3.3
tensorflow==2.4.0
termcolor==1.1.0
urllib3==1.26.2